backend/pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "Athlee"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "urllib3 (>=2.4.0,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "django-apscheduler (>=0.6.2,<1.0.0)",
    "label-studio-sdk (==1.0.10)",
    "boto3 (>=1.38.42,<2.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",  
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-v"
testpaths = [
    "accounts/tests",
    "common/tests",
    "group_users/tests",
    "projects/tests",
    "accounts",
    "common",
    "group_users",
    "projects"
]
norecursedirs = [
    "*/migrations",
    "*/static",
    "*/templates",
    "*/media",
    "*/staticfiles"
]

